+++ clients/admin-ui/cypress/e2e/config-wizard-plus.cy.ts
@@ -1,15 +1,18 @@
 import { stubPlus } from "cypress/support/stubs";
 
+import { ClusterHealth } from "~/types/api";
+
 /**
  * Handy function to bring us straight to the data flow scanner.
  * This could be put in a beforeEach, but then you can't overwrite intercepts
  */
 const goToDataFlowScanner = () => {
-  // Move past organization step.
-  cy.getByTestId("organization-info-form");
-  cy.getByTestId("submit-btn").click();
+  // Go through the initial config wizard steps
+  cy.visit("/add-systems");
+  cy.getByTestId("guided-setup-btn").click();
+  cy.wait("@getOrganization");
 
-  // Select data flow scanner to move to scan step.
+  // Select Runtime scanner to move to scan step.
   cy.getByTestId("add-system-form");
   cy.getByTestId("data-flow-scan-btn").click();
 };
@@ -28,27 +31,87 @@ describe("Config wizard with plus settings", () => {
     cy.intercept("PUT", "/api/v1/organization**", {
       fixture: "organization.json",
     }).as("updateOrganization");
-    stubPlus(true);
+  });
 
-    // Go through the initial config wizard steps
-    cy.visit("/add-systems");
-    cy.getByTestId("guided-setup-btn").click();
-    cy.wait("@getOrganization");
+  describe("Data flow scanner health", () => {
+    it("Disables data flow scanner button if it is not enabled", () => {
+      stubPlus(true, {
+        core_fidesctl_version: "1.9.6",
+        fidesctl_plus_server: "healthy",
+        fidescls_version: "1.0.3",
+        system_scanner: {
+          enabled: false,
+          cluster_health: null,
+          cluster_error: null,
+        },
+      });
+      cy.visit("/add-systems");
+      cy.getByTestId("guided-setup-btn").click();
+      cy.getByTestId("add-system-form");
+
+      cy.wait("@getPlusHealth");
+      cy.getByTestId("add-system-form");
+      cy.getByTestId("data-flow-scan-btn").should("be.disabled");
+      cy.getByTestId("cluster-health-indicator").should("not.exist");
+    });
+
+    it("Can show the scanner as unhealthy", () => {
+      stubPlus(true, {
+        core_fidesctl_version: "1.9.6",
+        fidesctl_plus_server: "healthy",
+        fidescls_version: "1.0.3",
+        system_scanner: {
+          enabled: true,
+          cluster_health: ClusterHealth.UNHEALTHY,
+          cluster_error: null,
+        },
+      });
+      cy.visit("/add-systems");
+      cy.getByTestId("guided-setup-btn").click();
+      cy.getByTestId("add-system-form");
+
+      cy.wait("@getPlusHealth");
+      cy.getByTestId("add-system-form");
+      cy.getByTestId("data-flow-scan-btn").should("be.disabled");
+      cy.getByTestId("cluster-health-indicator")
+        .invoke("attr", "title")
+        .should("eq", "Cluster is unhealthy");
+    });
+
+    it("Can show the scanner as enabled and healthy", () => {
+      stubPlus(true);
+      cy.visit("/add-systems");
+      cy.getByTestId("guided-setup-btn").click();
+      cy.getByTestId("add-system-form");
+
+      cy.wait("@getPlusHealth");
+      cy.getByTestId("add-system-form");
+      cy.getByTestId("data-flow-scan-btn").should("be.enabled");
+      cy.getByTestId("cluster-health-indicator")
+        .invoke("attr", "title")
+        .should("eq", "Cluster is connected and healthy");
+    });
   });
 
   describe("Data flow scanner steps", () => {
     beforeEach(() => {
+      stubPlus(true);
+
       // Stub scan endpoints
       cy.intercept("PUT", "/api/v1/plus/scan*", {
         delay: 500,
         fixture: "data-flow-scanner/list.json",
       }).as("putScanResults");
       cy.intercept("POST", "/api/v1/system/upsert", []).as("upsertSystems");
+      cy.intercept("GET", "/api/v1/plus/scan/latest?diff=true", {
+        statusCode: 404,
+        body: { detail: "No saved system scans found." },
+      }).as("getLatestScanDiff");
     });
 
     it("Allows calling the data flow scanner with classify", () => {
       cy.intercept("GET", "/api/v1/plus/classify?resource_type=systems*", {
-        fixture: "classify/list-systems.json",
+        fixture: "data-flow-scanner/list-classify-systems.json",
       }).as("getClassifyList");
 
       goToDataFlowScanner();
@@ -57,6 +120,7 @@ describe("Config wizard with plus settings", () => {
         const { url } = interception.request;
         expect(url).to.contain("classify=true");
       });
+      cy.wait("@getLatestScanDiff");
       cy.getByTestId("scan-results");
 
       // Check that the systems we expect are in the results table
@@ -121,6 +185,68 @@ describe("Config wizard with plus settings", () => {
         .then((rows) => {
           expect(rows.length).to.eql(2);
         });
+    });
+
+    it("Can rescan", () => {
+      let numAddedSystems = 0;
+      cy.fixture("data-flow-scanner/diff.json").then((diff) => {
+        cy.intercept("GET", "/api/v1/plus/scan/latest?diff=true", {
+          body: diff,
+        }).as("getLatestScanDiff");
+        numAddedSystems = diff.added_systems.length;
+      });
+      cy.intercept("GET", "/api/v1/plus/classify?resource_type=systems*", {
+        fixture: "classify/list-systems.json",
+      }).as("getClassifyList");
+
+      goToDataFlowScanner();
+      cy.getByTestId("scanner-loading");
+      cy.wait("@putScanResults");
+      cy.wait("@getLatestScanDiff");
+      cy.getByTestId("scan-results");
+
+      // Should render just the added systems discovered from the diff
+      cy.get("table")
+        .find("tbody > tr")
+        .then((rows) => {
+          expect(rows.length).to.eql(numAddedSystems);
+        });
+
+      cy.getByTestId("register-btn").click();
+      cy.wait("@upsertSystems");
+      cy.getByTestId("systems-classify-table")
+        .find("tbody > tr")
+        .then((rows) => {
+          expect(rows.length).to.eql(numAddedSystems);
+        });
+    });
+
+    it("Renders an empty state", () => {
+      // No newly added systems
+      cy.fixture("data-flow-scanner/diff.json").then((diff) => {
+        cy.intercept("GET", "/api/v1/plus/scan/latest?diff=true", {
+          body: { ...diff, added_systems: [] },
+        }).as("getLatestScanDiff");
+      });
+      goToDataFlowScanner();
+      cy.getByTestId("scanner-loading");
+      cy.wait("@putScanResults");
+      cy.wait("@getLatestScanDiff");
+      cy.getByTestId("scan-results");
+
+      // No results message
+      cy.getByTestId("no-results");
+      cy.getByTestId("back-btn").click();
+      cy.getByTestId("add-system-form");
+    });
+
+    it("Resets the flow when it is completed", () => {
+      goToDataFlowScanner();
+      cy.wait("@putScanResults");
+      cy.getByTestId("scan-results");
+      cy.getByTestId("register-btn").click();
+      cy.getByTestId("nav-link-Add Systems").click();
+      cy.getByTestId("setup");
     });
 
     it("Can render an error", () => {