+++ clients/admin-ui/src/features/plus/plus.slice.ts
@@ -1,28 +1,31 @@
 import { createSelector } from "@reduxjs/toolkit";
 import { createApi, fetchBaseQuery } from "@reduxjs/toolkit/query/react";
+import { addCommonHeaders } from "common/CommonHeaders";
 
-import { selectSystemsForReview } from "~/features/config-wizard/config-wizard.slice";
+import type { RootState } from "~/app/store";
+import { selectToken } from "~/features/auth";
 import {
   selectActiveCollection,
   selectActiveDatasetFidesKey,
   selectActiveField,
 } from "~/features/dataset/dataset.slice";
+import { selectSystemsToClassify } from "~/features/system";
 import {
   ClassificationResponse,
   ClassifyCollection,
   ClassifyDatasetResponse,
+  ClassifyEmpty,
   ClassifyField,
   ClassifyInstanceResponseValues,
   ClassifyRequestPayload,
   ClassifyStatusUpdatePayload,
+  ClassifySystem,
   GenerateTypes,
+  HealthCheck,
+  SystemScannerStatus,
   SystemScanResponse,
+  SystemsDiff,
 } from "~/types/api";
-
-interface HealthResponse {
-  core_fidesctl_version: string;
-  status: "healthy";
-}
 
 interface ScanParams {
   classify?: boolean;
@@ -37,10 +40,20 @@ export const plusApi = createApi({
   reducerPath: "plusApi",
   baseQuery: fetchBaseQuery({
     baseUrl: `${process.env.NEXT_PUBLIC_FIDESCTL_API}/plus`,
+    prepareHeaders: (headers, { getState }) => {
+      const token: string | null = selectToken(getState() as RootState);
+      addCommonHeaders(headers, token);
+      return headers;
+    },
   }),
-  tagTypes: ["Plus", "ClassifyInstancesDatasets", "ClassifyInstancesSystems"],
+  tagTypes: [
+    "Plus",
+    "ClassifyInstancesDatasets",
+    "ClassifyInstancesSystems",
+    "LatestScan",
+  ],
   endpoints: (build) => ({
-    getHealth: build.query<HealthResponse, void>({
+    getHealth: build.query<HealthCheck, void>({
       query: () => "health",
     }),
     /**
@@ -101,9 +114,18 @@ export const plusApi = createApi({
       },
     }),
     getClassifyDataset: build.query<ClassificationResponse, string>({
-      query: (dataset_fides_key: string) =>
-        `classify/details/${dataset_fides_key}`,
+      query: (dataset_fides_key: string) => ({
+        url: `classify/details/${dataset_fides_key}`,
+        params: { resource_type: GenerateTypes.DATASETS },
+      }),
       providesTags: ["ClassifyInstancesDatasets"],
+    }),
+    getClassifySystem: build.query<ClassifySystem | ClassifyEmpty, string>({
+      query: (fidesKey: string) => ({
+        url: `classify/details/${fidesKey}`,
+        params: { resource_type: GenerateTypes.SYSTEMS },
+      }),
+      providesTags: ["ClassifyInstancesSystems"],
     }),
 
     // Kubernetes Cluster Scanner
@@ -113,7 +135,15 @@ export const plusApi = createApi({
         params,
         method: "PUT",
       }),
-      invalidatesTags: ["ClassifyInstancesSystems"],
+      invalidatesTags: ["ClassifyInstancesSystems", "LatestScan"],
+    }),
+    getLatestScanDiff: build.query<SystemsDiff, void>({
+      query: () => ({
+        url: `scan/latest`,
+        params: { diff: true },
+        method: "GET",
+      }),
+      providesTags: ["LatestScan"],
     }),
   }),
 });
@@ -123,14 +153,27 @@ export const {
   useCreateClassifyInstanceMutation,
   useGetAllClassifyInstancesQuery,
   useGetClassifyDatasetQuery,
+  useGetClassifySystemQuery,
   useUpdateClassifyInstanceMutation,
   useUpdateScanMutation,
+  useGetLatestScanDiffQuery,
+  useLazyGetLatestScanDiffQuery,
 } = plusApi;
 
 export const useHasPlus = () => {
   const { isSuccess: hasPlus } = useGetHealthQuery();
   return hasPlus;
 };
+
+export const selectHealth: (state: RootState) => HealthCheck | undefined =
+  createSelector(plusApi.endpoints.getHealth.select(), ({ data }) => data);
+
+export const selectDataFlowScannerStatus: (
+  state: RootState
+) => SystemScannerStatus | undefined = createSelector(
+  plusApi.endpoints.getHealth.select(),
+  ({ data }) => data?.system_scanner
+);
 
 const emptyClassifyInstances: ClassifyInstanceResponseValues[] = [];
 export const selectDatasetClassifyInstances = createSelector(
@@ -141,11 +184,11 @@ export const selectDatasetClassifyInstances = createSelector(
 );
 
 export const selectSystemClassifyInstances = createSelector(
-  [(state) => state, selectSystemsForReview],
+  [(state) => state, selectSystemsToClassify],
   (state, systems) =>
     plusApi.endpoints.getAllClassifyInstances.select({
       resource_type: GenerateTypes.SYSTEMS,
-      fides_keys: systems.map((s) => s.fides_key),
+      fides_keys: systems?.map((s) => s.fides_key),
     })(state)?.data ?? emptyClassifyInstances
 );