+++ CHANGELOG.md
@@ -20,6 +20,7 @@ The types of changes are:
 ### Changed
 
 * The `celery.toml` is no longer used, instead it is a subsection of the `fides.toml` file [#1990](https://github.com/ethyca/fides/pull/1990)
+* The `fideslib` module has been merged into `fides`, code redundancies have been removed [#1859](https://github.com/ethyca/fides/pull/1859)
 * Update sample project landing page copy to be version-agnostic [#1958](https://github.com/ethyca/fides/pull/1958)
 
 ## [2.2.0](https://github.com/ethyca/fides/compare/2.1.0...2.2.0)